requirements.txt
streamlit==1.32.0
groq==0.4.2
PyMuPDF==1.23.8
python-docx==0.8.11
numpy==1.24.3
python-dotenv==1.0.0
scikit-learn==1.3.2
